requirements.txt
jinja2~=3.0
pydantic~=1.8
pyyaml~=6.0
aioregistry~=0.3
